REFACTOR: Move jest-util to test directory and update imports
- Relocate jest-util.ts from src/util/ to test/ directory
- Add @test/* path alias to tsconfig.json
- Update import paths in spec files from @src/util/jest-util to @test/jest-util
